CD-458 fixes

+++ Bundles/Touch/src/SprykerEngine/Zed/Touch/Business/Model/TouchRecord.php
@@ -120,8 +120,9 @@ class TouchRecord implements TouchRecordInterface
             $touchActiveEntity = $this->touchQueryContainer->queryUpdateTouchEntry($itemType, $idItem, SpyTouchTableMap::COL_ITEM_EVENT_ACTIVE)
                     ->findOne()
                 ;
-            $touchActiveEntity->delete();
-
+            if (null !== $touchActiveEntity) {
+                $touchActiveEntity->delete();
+            }
             return true;
         }
